Move end-to-end/rot13 test to Mocha and Chai Install Mocha and Chai
