+++ package.json
@@ -37,6 +37,8 @@
   },
   "devDependencies": {
     "ava": "^0.17.0",
+    "chai": "^3.5.0",
+    "mocha": "^3.2.0",
     "nyc": "^10.0.0",
     "xo": "^0.17.0"
   },
