+++ test/end-to-end/rot13.js
@@ -1,0 +1,40 @@
+const assert = require('chai').assert;
+const brunfaick = require('../../.');
+
+describe('end-to-end > rot13.b', () => {
+  it('should outputs its code', () => {
+    const program = `
+    ,
+    [>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
+    [>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
+    [>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
+    [>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
+    [>++++++++++++++<-
+    [>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
+    [>>+++++[<----->-]<<-
+    [>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
+    [>++++++++++++++<-
+    [>+<-[>+<-[>+<-[>+<-[>+<-
+    [>++++++++++++++<-
+    [>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
+    [>>+++++[<----->-]<<-
+    [>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
+    [>++++++++++++++<-
+    [>+<-]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]
+    ]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]>.[-]<,]
+
+    of course any function char f(char) can be made easily on the same principle
+
+    [Daniel B Cristofani (cristofdathevanetdotcom)
+    http://www.hevanet.com/cristofd/brainfuck/]
+    `;
+
+    const input = 'foo bar baz';
+
+    const result = brunfaick(program, input);
+
+    const expected = 'sbb one onm';
+
+    assert.strictEqual(result, expected);
+  });
+});
