--- test/performance/rot13.js
@@ -1,38 +1,0 @@
-const test = require('ava');
-const brunfaick = require('../../.');
-
-test('Outputs its code.', t => {
-  const program = `
-	,
-	[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>++++++++++++++<-
-	[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>>+++++[<----->-]<<-
-	[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>++++++++++++++<-
-	[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>++++++++++++++<-
-	[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>>+++++[<----->-]<<-
-	[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-[>+<-
-	[>++++++++++++++<-
-	[>+<-]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]
-	]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]>.[-]<,]
-
-	of course any function char f(char) can be made easily on the same principle
-
-	[Daniel B Cristofani (cristofdathevanetdotcom)
-	http://www.hevanet.com/cristofd/brainfuck/]
-  `;
-
-  const input = 'foo bar baz';
-
-  const result = brunfaick(program, input);
-
-  const expected = 'sbb one onm';
-
-  t.is(result, expected);
-});
